requirements.txt
antlr4-python3-runtime==4.8
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
configparser==5.1.0
docker-pycreds==0.4.0
gitdb==4.0.9
GitPython==3.1.24
hydra-core==1.1.1
idna==3.3
mkl-fft
mkl-random
mkl-service==2.4.0
numpy
olefile
omegaconf==2.1.1
pathtools==0.1.2
Pillow==8.4.0
promise==2.3
protobuf==3.19.1
psutil==5.8.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.26.0
sentry-sdk==1.4.3
shortuuid==1.0.8
six
smmap==5.0.0
subprocess32==3.5.4
termcolor==1.1.0
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.1
tqdm==4.62.3
typing-extensions
urllib3==1.26.7
wandb==0.12.6
yaspin==2.1.0
